backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
GitPython
langchain
langchain-community
langchain-text-splitters
langchain-openai
supabase
tiktoken
